MongoDB Integrated

+++ src/main/java/com/practice/controller/UserController.java
@@ -36,7 +36,7 @@ public class UserController {
     }
 
     @DeleteMapping("/{id}")
-    public String deleteUser(@PathVariable("id") int id){
+    public String deleteUser(@PathVariable("id") Long id){
         userService.deleteUser(id);
         return "redirect:/user";
     }
